rename page components and move them to pages directory
