--- react_ui/src/components/About/index.ts
@@ -1,2 +1,0 @@
-import About from "./About";
-export default About;
