--- react_ui/src/components/Contact/index.ts
@@ -1,2 +1,0 @@
-import Contact from "./Contact";
-export default Contact;
