--- react_ui/src/components/Home/Home.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import AnimatedPage from '../AnimatedPage';
-
-const Home = () => {
-  return (
-    <AnimatedPage>
-      Welcome!
-    </AnimatedPage>
-  )
-};
-
-export default Home;
