--- react_ui/src/components/Home/index.ts
@@ -1,2 +1,0 @@
-import Home from "./Home";
-export default Home;
