--- react_ui/src/components/Notes/index.ts
@@ -1,2 +1,0 @@
-import Notes from "./Notes";
-export default Notes;
