--- react_ui/src/components/Resume/index.ts
@@ -1,2 +1,0 @@
-import Resume from "./Resume";
-export default Resume;
