+++ react_ui/src/pages/HomePage.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import AnimatedPage from '../components/AnimatedPage';
+
+const HomePage = () => {
+  return (
+    <AnimatedPage>
+      Welcome!
+    </AnimatedPage>
+  )
+};
+
+export default HomePage;
